Backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
aspose-words==25.5.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cssselect==1.3.0
cycler==0.12.1
distro==1.9.0
docx==0.2.4
docx2pdf==0.1.8
docx2python==3.5.0
ecdsa==0.19.1
fastapi==0.115.12
fonttools==4.58.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.3
numpy==2.2.6
openai==1.77.0
packaging==25.0
paragraphs==1.0.1
passlib==1.7.4
pdf2image==1.17.0
pillow==11.2.1
pyasn1==0.4.8
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docs==0.1.0
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pywin32==310
reportlab==4.4.1
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
tqdm==4.67.1
types-html5lib==1.1.11.20250516
types-lxml==2025.3.30
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
